test(asr): add multilingual ASR integration tests
