+++ apps/ml/tests/test_asr.py
@@ -221,54 +221,3 @@ def test_health_endpoint():
     assert response.json()["status"] == "healthy"
 
 
-# ── 6. Real language audio fixtures ──────────────────────────────────────────
-# These tests are skipped automatically if fixture files are missing.
-# Run locally after downloading real audio samples.
-
-@pytest.mark.skipif(
-    not os.path.exists(os.path.join(FIXTURES_DIR, "hindi_sample.wav"))
-    or shutil.which("ffmpeg") is None,
-    reason="Hindi fixture not found or ffmpeg not available",
-)
-def test_hindi_language_detection():
-    """Real Hindi audio must be detected as 'hi' or 'ur' (Whisper limitation)."""
-    with open(os.path.join(FIXTURES_DIR, "hindi_sample.wav"), "rb") as f:
-        response = client.post(
-            "/asr/transcribe",
-            files={"file": ("hindi_sample.wav", f, "audio/wav")},
-        )
-    assert response.status_code == 200
-    assert response.json()["language"] in ["hi", "ur"], \
-        f"Expected hi or ur, got: {response.json()['language']}"
-
-
-@pytest.mark.skipif(
-    not os.path.exists(os.path.join(FIXTURES_DIR, "tamil_sample.wav"))
-    or shutil.which("ffmpeg") is None,
-    reason="Tamil fixture not found or ffmpeg not available",
-)
-def test_tamil_language_detection():
-    """Real Tamil audio must be detected as 'ta'."""
-    with open(os.path.join(FIXTURES_DIR, "tamil_sample.wav"), "rb") as f:
-        response = client.post(
-            "/asr/transcribe",
-            files={"file": ("tamil_sample.wav", f, "audio/wav")},
-        )
-    assert response.status_code == 200
-    assert response.json()["language"] == "ta"
-
-
-@pytest.mark.skipif(
-    not os.path.exists(os.path.join(FIXTURES_DIR, "bengali_sample.wav"))
-    or shutil.which("ffmpeg") is None,
-    reason="Bengali fixture not found or ffmpeg not available",
-)
-def test_bengali_language_detection():
-    """Real Bengali audio must be detected as 'bn'."""
-    with open(os.path.join(FIXTURES_DIR, "bengali_sample.wav"), "rb") as f:
-        response = client.post(
-            "/asr/transcribe",
-            files={"file": ("bengali_sample.wav", f, "audio/wav")},
-        )
-    assert response.status_code == 200
-    assert response.json()["language"] == "bn"
